TI thermal framework: omap die governor: simplify the cooling device search
TI thermal framework: omap die governor: simplify the cooling device search

Inside each zone entering function, there is a search for cooling devices.
This search is not really needed as the registered devices are already
supposed to be part of that thermal zone. The checks for the missing pointers
are done by the caller macro helper.

Change-Id: I4a227933f3b7c723675007c5b7dc7d3ce1ec1ac6
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>